AI-Enhanced Drone Imaginary Pipeline for Crop Protection Field Trails/requirements.txt
# Core
streamlit==1.41.0
plotly==5.24.1
folium==0.18.0
streamlit-folium==0.23.1
branca==0.8.1

# Geospatial
rasterio==1.4.3
numpy==1.26.4
pandas==2.2.3
geopandas==1.0.1
shapely==2.0.6
pyproj==3.7.0

# Image Processing
Pillow==11.1.0
scikit-image==0.24.0
scipy==1.14.1
matplotlib==3.9.3

# API & Networking
requests==2.32.3
python-dotenv==1.0.1

# Scheduling
schedule==1.2.2
watchdog==6.0.0
